LayoutTests/platform/gtk/accessibility/text-for-range-heading-expected.txt

A text header

Block span in a link in a heading
Inline span in a link in a heading

Block span in a heading
Inline span in a heading

This tests the ability to get element text for a specified range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS heading1.stringForRange(0, -1) is 'A text header'
PASS heading2.stringForRange(0, -1) is 'Block span in a link in a heading\nInline span in a link in a heading'
PASS heading3.stringForRange(0, -1) is 'Block span in a heading\nInline span in a heading'
PASS successfullyParsed is true

TEST COMPLETE
